+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to the vscode-chat extension will be documented in this file
 ### Added
 
 - Support for muted channels: new messages in muted channels do not update the unread count.
-- Added support to expand thread messages to render replies.
+- Added support to expand thread messages to render replies, with text and file attachments.
 - Consistency between Slack clients for unread counts: marking messages as read from other Slack clients will update the unread count.
 - Unread messages are now highlighted in the webview.
 - Added documentation for [product roadmap](VISION.md) and [adding chat providers](PROVIDERS.md).